Created common conftest with fixtures useful in other tests
