+++ lambdas/drain_ecs_container_instance/conftest.py
@@ -1,0 +1,1 @@
+../common/conftest.py
