+++ lambdas/drain_ecs_container_instance/test_drain_ecs_container_instance.py
@@ -2,27 +2,11 @@ import json
 
 import boto3
 from mock import Mock
-from moto import mock_ec2, mock_autoscaling, mock_ecs, mock_sns, mock_sqs
 from moto.ec2 import utils as ec2_utils
 
 import drain_ecs_container_instance
 
 import pytest
-
-
-@pytest.fixture()
-def moto_start():
-    mock_autoscaling().start()
-    mock_ec2().start()
-    mock_ecs().start()
-    mock_sns().start()
-    mock_sqs().start()
-    yield
-    mock_autoscaling().stop()
-    mock_ec2().stop()
-    mock_ecs().stop()
-    mock_sns().stop()
-    mock_sqs().stop()
 
 
 @pytest.fixture()
@@ -45,28 +29,6 @@ def autoscaling_group(moto_start):
     instances = fake_ec2_client.describe_instances()
     instance_id = instances['Reservations'][0]['Instances'][0]['InstanceId']
     yield auto_scaling_group_name, instance_id
-
-
-@pytest.fixture()
-def sns_sqs(moto_start):
-    fake_sns_client = boto3.client('sns')
-    fake_sqs_client = boto3.client('sqs')
-    queue_name = "test-queue"
-    topic_name = "test-topic"
-    print(f"Creating topic {topic_name} and queue {queue_name}")
-
-    fake_sns_client.create_topic(Name=topic_name)
-    response = fake_sns_client.list_topics()
-    topic_arn = response["Topics"][0]['TopicArn']
-
-    queue = fake_sqs_client.create_queue(QueueName=queue_name)
-
-    fake_sns_client.subscribe(
-        TopicArn=topic_arn,
-        Protocol="sqs",
-        Endpoint=f"arn:aws:sqs:eu-west-1:123456789012:{queue_name}"
-    )
-    yield topic_arn, queue['QueueUrl']
 
 
 @pytest.fixture()
